Add combat package Refactor damageindicator to be in combat
