+++ src/main/java/at/hannibal2/skyhanni/events/DamageIndicatorDetectedEvent.kt
@@ -1,5 +1,5 @@
 package at.hannibal2.skyhanni.events
 
-import at.hannibal2.skyhanni.features.damageindicator.EntityData
+import at.hannibal2.skyhanni.features.combat.damageindicator.EntityData
 
 class DamageIndicatorDetectedEvent(val entityData: EntityData) : LorenzEvent()
