+++ src/main/java/at/hannibal2/skyhanni/features/combat/damageindicator/DamageCounter.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.features.damageindicator
+package at.hannibal2.skyhanni.features.combat.damageindicator
 
 import java.util.LinkedList
