+++ src/main/java/at/hannibal2/skyhanni/features/combat/damageindicator/OldDamage.kt
@@ -1,3 +1,3 @@
-package at.hannibal2.skyhanni.features.damageindicator
+package at.hannibal2.skyhanni.features.combat.damageindicator
 
 class OldDamage(val time: Long, val damage: Long, val healing: Long)
